huggingface/hf/requirements.txt
streamlit
openai
pinecone==3.0.2
langchain>=0.1.13
langchain-community>=0.0.20
tiktoken
Pillow